+++ src/libromdata/Other/ISO.cpp
@@ -3,12 +3,14 @@
  * ISO.cpp: ISO-9660 disc image parser.                                    *
  *                                                                         *
  * Copyright (c) 2019-2020 by David Korth.                                 *
+ * Copyright (c) 2020 by Egor.                                             *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
 #include "stdafx.h"
 #include "ISO.hpp"
 #include "../iso_structs.h"
+#include "hsfs_structs.h"
 
 // librpbase, librpfile, librpcpu
 #include "common.h"
@@ -44,8 +46,22 @@ class ISOPrivate : public LibRpBase::RomDataPrivate
 		RP_DISABLE_COPY(ISOPrivate)
 
 	public:
-		// ISO primary volume descriptor.
-		ISO_Primary_Volume_Descriptor pvd;
+		// Disc type.
+		enum class DiscType {
+			Unknown = -1,
+
+			ISO9660 = 0,
+			HighSierra = 1,
+
+			Max
+		};
+		DiscType discType;
+
+		// Primary volume descriptor.
+		union {
+			ISO_Primary_Volume_Descriptor iso;	// ISO-9660
+			HSFS_Primary_Volume_Descriptor hsfs;	// High Sierra
+		} pvd;
 
 		// Sector size.
 		// Usually 2048 or 2352.
@@ -76,12 +92,45 @@ class ISOPrivate : public LibRpBase::RomDataPrivate
 			// which doesn't take an ISO_PVD_DateTime_t struct.
 			return RomDataPrivate::pvd_time_to_unix_time(pvd_time->full, pvd_time->tz_offset);
 		}
+
+		/**
+		 * Convert an HSFS PVD timestamp to UNIX time.
+		 * @param pvd_time PVD timestamp
+		 * @return UNIX time, or -1 if invalid or not set.
+		 */
+		static inline time_t pvd_time_to_unix_time(const HSFS_PVD_DateTime_t *pvd_time)
+		{
+			// Wrapper for RomData::pvd_time_to_unix_time(),
+			// which doesn't take an HSFS_PVD_DateTime_t struct.
+			return RomDataPrivate::pvd_time_to_unix_time(pvd_time->full, 0);
+		}
+
+		/**
+		 * Add fields common to HSFS and ISO-9660 (except timestamps)
+		 * @param pvd PVD
+		 */
+		template<typename T>
+		void addPVDCommon(const T *pvd);
+
+		/**
+		 * Add timestamp fields from PVD
+		 * @param pvd PVD
+		 */
+		template<typename T>
+		void addPVDTimestamps(const T *pvd);
+
+		/**
+		 * Check the PVD and determine its type.
+		 * @return DiscType value. (DiscType::Unknown if not valid)
+		 */
+		DiscType checkPVD(void) const;
 };
 
 /** ISOPrivate **/
 
 ISOPrivate::ISOPrivate(ISO *q, IRpFile *file)
 	: super(q, file)
+	, discType(DiscType::Unknown)
 	, sector_size(0)
 	, sector_offset(0)
 	, s_udf_version(nullptr)
@@ -178,6 +227,133 @@ void ISOPrivate::checkVolumeDescriptors(void)
 	// TODO: More descriptors?
 }
 
+/**
+ * Add fields common to HSFS and ISO-9660 (except timestamps)
+ * @param pvd PVD
+ */
+template<typename T>
+void ISOPrivate::addPVDCommon(const T *pvd)
+{
+	// System ID
+	fields->addField_string(C_("ISO", "System ID"),
+		latin1_to_utf8(pvd->sysID, sizeof(pvd->sysID)),
+		RomFields::STRF_TRIM_END);
+
+	// Volume ID
+	fields->addField_string(C_("ISO", "Volume ID"),
+		latin1_to_utf8(pvd->volID, sizeof(pvd->volID)),
+		RomFields::STRF_TRIM_END);
+
+	// Size of volume
+	fields->addField_string(C_("ISO", "Volume Size"),
+		formatFileSize(
+			static_cast<off64_t>(pvd->volume_space_size.he) *
+			static_cast<off64_t>(pvd->logical_block_size.he)));
+
+	// TODO: Show block size?
+
+	// Disc number
+	if (pvd->volume_seq_number.he != 0 && pvd->volume_set_size.he > 1) {
+		const char *const disc_number_title = C_("RomData", "Disc #");
+		fields->addField_string(disc_number_title,
+			// tr: Disc X of Y (for multi-disc games)
+			rp_sprintf_p(C_("RomData|Disc", "%1$u of %2$u"),
+				pvd->volume_seq_number.he,
+				pvd->volume_set_size.he));
+	}
+
+	// Volume set ID
+	fields->addField_string(C_("ISO", "Volume Set"),
+		latin1_to_utf8(pvd->volume_set_id, sizeof(pvd->volume_set_id)),
+		RomFields::STRF_TRIM_END);
+
+	// Publisher
+	fields->addField_string(C_("ISO", "Publisher"),
+		latin1_to_utf8(pvd->publisher, sizeof(pvd->publisher)),
+		RomFields::STRF_TRIM_END);
+
+	// Data Preparer
+	fields->addField_string(C_("ISO", "Data Preparer"),
+		latin1_to_utf8(pvd->data_preparer, sizeof(pvd->data_preparer)),
+		RomFields::STRF_TRIM_END);
+
+	// Application
+	fields->addField_string(C_("ISO", "Application"),
+		latin1_to_utf8(pvd->application, sizeof(pvd->application)),
+		RomFields::STRF_TRIM_END);
+
+	// Copyright file
+	fields->addField_string(C_("ISO", "Copyright File"),
+		latin1_to_utf8(pvd->copyright_file, sizeof(pvd->copyright_file)),
+		RomFields::STRF_TRIM_END);
+
+	// Abstract file
+	fields->addField_string(C_("ISO", "Abstract File"),
+		latin1_to_utf8(pvd->abstract_file, sizeof(pvd->abstract_file)),
+		RomFields::STRF_TRIM_END);
+}
+
+/**
+ * Add timestamp fields from PVD
+ * @param pvd PVD
+ */
+template<typename T>
+void ISOPrivate::addPVDTimestamps(const T *pvd)
+{
+	// TODO: Show the original timezone?
+	// For now, converting to UTC and showing as local time.
+
+	// Volume creation time
+	fields->addField_dateTime(C_("ISO", "Creation Time"),
+		pvd_time_to_unix_time(&pvd->btime),
+		RomFields::RFT_DATETIME_HAS_DATE |
+		RomFields::RFT_DATETIME_HAS_TIME);
+
+	// Volume modification time
+	fields->addField_dateTime(C_("ISO", "Modification Time"),
+		pvd_time_to_unix_time(&pvd->mtime),
+		RomFields::RFT_DATETIME_HAS_DATE |
+		RomFields::RFT_DATETIME_HAS_TIME);
+
+	// Volume expiration time
+	fields->addField_dateTime(C_("ISO", "Expiration Time"),
+		pvd_time_to_unix_time(&pvd->exptime),
+		RomFields::RFT_DATETIME_HAS_DATE |
+		RomFields::RFT_DATETIME_HAS_TIME);
+
+	// Volume effective time
+	fields->addField_dateTime(C_("ISO", "Effective Time"),
+		pvd_time_to_unix_time(&pvd->efftime),
+		RomFields::RFT_DATETIME_HAS_DATE |
+		RomFields::RFT_DATETIME_HAS_TIME);
+}
+
+/**
+ * Check the PVD and determine its type.
+ * @return DiscType value. (DiscType::Unknown if not valid)
+ */
+ISOPrivate::DiscType ISOPrivate::checkPVD(void) const
+{
+	if (pvd.iso.header.type == ISO_VDT_PRIMARY &&
+	    pvd.iso.header.version == ISO_VD_VERSION &&
+	    !memcmp(pvd.iso.header.identifier, ISO_VD_MAGIC, sizeof(pvd.iso.header.identifier)))
+	{
+		// ISO-9660 PVD.
+		return DiscType::ISO9660;
+	}
+
+	if (pvd.hsfs.header.type == ISO_VDT_PRIMARY &&
+	    pvd.hsfs.header.version == HSFS_VD_VERSION &&
+	    !memcmp(pvd.hsfs.header.identifier, HSFS_VD_MAGIC, sizeof(pvd.hsfs.header.identifier)))
+	{
+		// High Sierra PVD.
+		return DiscType::HighSierra;
+	}
+
+	// Not recognized.
+	return DiscType::Unknown;
+}
+
 /** ISO **/
 
 /**
@@ -220,10 +396,8 @@ ISO::ISO(IRpFile *file)
 	// Check if the PVD is valid.
 	// NOTE: Not using isRomSupported_static(), since this function
 	// only checks the file extension.
-	if (d->pvd.header.type == ISO_VDT_PRIMARY &&
-	    d->pvd.header.version == ISO_VD_VERSION &&
-	    !memcmp(d->pvd.header.identifier, ISO_VD_MAGIC, sizeof(d->pvd.header.identifier)))
-	{
+	d->discType = d->checkPVD();
+	if (d->discType > ISOPrivate::DiscType::Unknown) {
 		// Found the PVD using 2048-byte sectors.
 		d->sector_size = ISO_SECTOR_SIZE_MODE1_COOKED;
 		d->sector_offset = ISO_DATA_OFFSET_MODE1_COOKED;
@@ -238,10 +412,8 @@ ISO::ISO(IRpFile *file)
 			return;
 		}
 
-		if (d->pvd.header.type == ISO_VDT_PRIMARY &&
-		    d->pvd.header.version == ISO_VD_VERSION &&
-		    !memcmp(d->pvd.header.identifier, ISO_VD_MAGIC, sizeof(d->pvd.header.identifier)))
-		{
+		d->discType = d->checkPVD();
+		if (d->discType > ISOPrivate::DiscType::Unknown) {
 			// Found the PVD using 2352-byte sectors.
 			d->sector_size = ISO_SECTOR_SIZE_MODE1_RAW;
 			d->sector_offset = ISO_DATA_OFFSET_MODE1_RAW;
@@ -257,7 +429,9 @@ ISO::ISO(IRpFile *file)
 	d->isValid = true;
 
 	// Check for additional volume descriptors.
-	d->checkVolumeDescriptors();
+	if (d->discType == ISOPrivate::DiscType::ISO9660) {
+		d->checkVolumeDescriptors();
+	}
 }
 
 /** ROM detection functions. **/
@@ -306,11 +480,16 @@ const char *ISO::systemName(unsigned int type) const
 		"ISO::systemName() array index optimization needs to be updated.");
 
 	// TODO: UDF, HFS, others?
-	static const char *const sysNames[4] = {
-		"ISO-9660", "ISO", "ISO", nullptr
+	static const char *const sysNames[2][4] = {
+		{"ISO-9660", "ISO", "ISO", nullptr},
+		{"High Sierra Format", "High Sierra", "HSF", nullptr},
 	};
 
-	return sysNames[type & SYSNAME_TYPE_MASK];
+	unsigned int sysID = 0;
+	if (d->discType == ISOPrivate::DiscType::HighSierra) {
+		sysID = 1;
+	}
+	return sysNames[sysID][type & SYSNAME_TYPE_MASK];
 }
 
 /**
@@ -334,6 +513,7 @@ const char *const *ISO::supportedFileExtensions_static(void)
 		".bin",		// BIN (2352-byte)
 		".xiso",	// Xbox ISO image
 		// TODO: More?
+		// TODO: Is there a separate extension for High Sierra?
 
 		nullptr
 	};
@@ -358,6 +538,7 @@ const char *const *ISO::supportedMimeTypes_static(void)
 		"application/x-iso9660-image",
 
 		// TODO: BIN (2352)?
+		// TODO: Is there a separate MIME for High Sierra?
 		nullptr
 	};
 	return mimeTypes;
@@ -382,107 +563,45 @@ int ISO::loadFieldData(void)
 		return -EIO;
 	}
 
-	// ISO-9660 Primary Volume Descriptor.
-	// TODO: Other descriptors?
-	const ISO_Primary_Volume_Descriptor *const pvd = &d->pvd;
 	d->fields->reserve(16);	// Maximum of 16 fields.
 
 	// NOTE: All fields are space-padded. (0x20, ' ')
 	// TODO: ascii_to_utf8()?
 
-	// ISO-9660 PVD
-	d->fields->setTabName(0, C_("ISO", "ISO-9660 PVD"));
+	switch (d->discType) {
+		case ISOPrivate::DiscType::ISO9660:
+			// ISO-9660
+			d->fields->setTabName(0, C_("ISO", "ISO-9660 PVD"));
 
-	// System ID
-	d->fields->addField_string(C_("ISO", "System ID"),
-		latin1_to_utf8(pvd->sysID, sizeof(pvd->sysID)),
-		RomFields::STRF_TRIM_END);
-	
-	// Volume ID
-	d->fields->addField_string(C_("ISO", "Volume ID"),
-		latin1_to_utf8(pvd->volID, sizeof(pvd->volID)),
-		RomFields::STRF_TRIM_END);
+			// PVD common fields (ISO-9660, High Sierra)
+			d->addPVDCommon(&d->pvd.iso);
 
-	// Size of volume
-	d->fields->addField_string(C_("ISO", "Volume Size"),
-		formatFileSize(
-			static_cast<off64_t>(pvd->volume_space_size.he) *
-			static_cast<off64_t>(pvd->logical_block_size.he)));
+			// Bibliographic file
+			d->fields->addField_string(C_("ISO", "Bibliographic File"),
+				latin1_to_utf8(d->pvd.iso.bibliographic_file, sizeof(d->pvd.iso.bibliographic_file)),
+				RomFields::STRF_TRIM_END);
 
-	// TODO: Show block size?
+			// Timestamps
+			d->addPVDTimestamps(&d->pvd.iso);
+			break;
 
-	// Disc number
-	if (pvd->volume_seq_number.he != 0 && pvd->volume_set_size.he > 1) {
-		const char *const disc_number_title = C_("RomData", "Disc #");
-		d->fields->addField_string(disc_number_title,
-			// tr: Disc X of Y (for multi-disc games)
-			rp_sprintf_p(C_("RomData|Disc", "%1$u of %2$u"),
-				pvd->volume_seq_number.he,
-				pvd->volume_set_size.he));
+		case ISOPrivate::DiscType::HighSierra:
+			// High Sierra
+			d->fields->setTabName(0, C_("ISO", "High Sierra PVD"));
+
+			// PVD common fields (ISO-9660, High Sierra)
+			d->addPVDCommon(&d->pvd.hsfs);
+
+			// Timestamps
+			d->addPVDTimestamps(&d->pvd.hsfs);
+			break;
+
+		default:
+			// Should not get here...
+			assert(!"Invalid ISO disc type.");
+			d->fields->setTabName(0, "ISO");
+			break;
 	}
-
-	// Volume set ID
-	d->fields->addField_string(C_("ISO", "Volume Set"),
-		latin1_to_utf8(pvd->volume_set_id, sizeof(pvd->volume_set_id)),
-		RomFields::STRF_TRIM_END);
-
-	// Publisher
-	d->fields->addField_string(C_("ISO", "Publisher"),
-		latin1_to_utf8(pvd->publisher, sizeof(pvd->publisher)),
-		RomFields::STRF_TRIM_END);
-
-	// Data Preparer
-	d->fields->addField_string(C_("ISO", "Data Preparer"),
-		latin1_to_utf8(pvd->data_preparer, sizeof(pvd->data_preparer)),
-		RomFields::STRF_TRIM_END);
-
-	// Application
-	d->fields->addField_string(C_("ISO", "Application"),
-		latin1_to_utf8(pvd->application, sizeof(pvd->application)),
-		RomFields::STRF_TRIM_END);
-
-	// Copyright file
-	d->fields->addField_string(C_("ISO", "Copyright File"),
-		latin1_to_utf8(pvd->copyright_file, sizeof(pvd->copyright_file)),
-		RomFields::STRF_TRIM_END);
-
-	// Abstract file
-	d->fields->addField_string(C_("ISO", "Abstract File"),
-		latin1_to_utf8(pvd->abstract_file, sizeof(pvd->abstract_file)),
-		RomFields::STRF_TRIM_END);
-
-	// Bibliographic file
-	d->fields->addField_string(C_("ISO", "Bibliographic File"),
-		latin1_to_utf8(pvd->bibliographic_file, sizeof(pvd->bibliographic_file)),
-		RomFields::STRF_TRIM_END);
-
-	/** Timestamps **/
-	// TODO: Show the original timezone?
-	// For now, converting to UTC and showing as local time.
-
-	// Volume creation time
-	d->fields->addField_dateTime(C_("ISO", "Creation Time"),
-		d->pvd_time_to_unix_time(&pvd->btime),
-		RomFields::RFT_DATETIME_HAS_DATE |
-		RomFields::RFT_DATETIME_HAS_TIME);
-
-	// Volume modification time
-	d->fields->addField_dateTime(C_("ISO", "Modification Time"),
-		d->pvd_time_to_unix_time(&pvd->mtime),
-		RomFields::RFT_DATETIME_HAS_DATE |
-		RomFields::RFT_DATETIME_HAS_TIME);
-
-	// Volume expiration time
-	d->fields->addField_dateTime(C_("ISO", "Expiration Time"),
-		d->pvd_time_to_unix_time(&pvd->exptime),
-		RomFields::RFT_DATETIME_HAS_DATE |
-		RomFields::RFT_DATETIME_HAS_TIME);
-
-	// Volume effective time
-	d->fields->addField_dateTime(C_("ISO", "Effective Time"),
-		d->pvd_time_to_unix_time(&pvd->efftime),
-		RomFields::RFT_DATETIME_HAS_DATE |
-		RomFields::RFT_DATETIME_HAS_TIME);
 
 	if (d->s_udf_version) {
 		// UDF version.